+++ dart/geocore/lib/src/base/temporal.dart
@@ -1,2 +1,3 @@
 export 'temporal/instant.dart';
 export 'temporal/interval.dart';
+export 'temporal/temporal.dart';